+++ src/repository/repository.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Transaction } from '../transactions/entities/transaction.entity';
+import { TransactionsRepository } from './transactions.repository';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Transaction])],
+  providers: [TransactionsRepository],
+  exports: [TransactionsRepository],
+})
+export class RepositoryModule {}
