+++ src/transactions/transactions.module.ts
@@ -2,9 +2,13 @@ import { BullModule } from '@nestjs/bull';
 import { Module } from '@nestjs/common';
 import { TransactionsController } from './transactions.controller';
 import { TransactionsService } from './transactions.service';
+import { RepositoryModule } from '../repository/repository.module';
 
 @Module({
-  imports: [BullModule.registerQueue({ name: 'transactions' })],
+  imports: [
+    BullModule.registerQueue({ name: 'transactions' }),
+    RepositoryModule,
+  ],
   controllers: [TransactionsController],
   providers: [TransactionsService],
 })
