+++ src/workers/transactions.worker.ts
@@ -2,7 +2,7 @@ import { Process, Processor } from '@nestjs/bull';
 import { ConflictException, Inject, Logger } from '@nestjs/common';
 import type { Job } from 'bull';
 import type { CreateTransactionDto } from '../transactions/dto/create-transaction.dto';
-import { TransactionsRepository } from './transactions.repository';
+import { TransactionsRepository } from '../repository/transactions.repository';
 
 @Processor('transactions')
 export class TransactionsWorker {
